Fix JVM tests and Windows build by upgrading kotlinx-datetime to 0.7.1 (#29)
Closes issue #19

---------

+++ composeApp/src/commonMain/kotlin/com/klemfner/whoscalling/ui/common/utils/TimeGrouping.kt
@@ -1,7 +1,7 @@
 package com.klemfner.whoscalling.ui.common.utils
 
-import kotlinx.datetime.Clock
-import kotlinx.datetime.Instant
+import kotlin.time.Clock
+import kotlin.time.Instant
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.toLocalDateTime
 
@@ -14,7 +14,7 @@ fun getTimePeriod(timestampMillis: Long): TimePeriod {
     val now = Clock.System.now().toLocalDateTime(tz).date
     val date = Instant.fromEpochMilliseconds(timestampMillis).toLocalDateTime(tz).date
 
-    val daysDiff = now.toEpochDays() - date.toEpochDays()
+    val daysDiff = (now.toEpochDays() - date.toEpochDays()).toInt()
 
     return when {
         daysDiff == 0 -> TimePeriod.TODAY
